requirements.txt
httpx
python-dotenv